+++ src/types/index.ts
@@ -1,15 +1,19 @@
-import { TOffer, TLocation, TOfferList, TCityName } from './offer';
+import { TOffer, TOfferList, TOfferLight } from './offer';
 import { TUser, TUserAuthorisation } from './user';
 import { TReview, TReviewList } from './review';
+import { TLocation } from './location';
+import { TCity, TCityName } from './city';
 
 export type {
   TOffer,
   TOfferList,
+  TOfferLight,
   TLocation,
   TUser,
   TUserAuthorisation,
   TReview,
   TReviewList,
   TCityName,
+  TCity,
 };
