+++ src/types/location.ts
@@ -1,0 +1,5 @@
+export type TLocation = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+}
